Various changes
Change List
-Added validation.php it provides functions to validate name, email, and
address. The address validation depends on the Lob api which is
everything in the vendor folder. The Lob api validates real U.S and
international addresses so we will probably have to change our records
to reflect that.
-Account page more cleanup in terms of visuals. I changed the account
page so that it hides the billingid when attempting to edit a payment
method.
-Edit_info page added in the validation and removed country
-Edit payment my first commit of this page. Right now it has no
validation because we are storing payment info ourselves will probably
add some simple validation (ex. all numbers and appropriate length for
card number)
-Inventory page changes from reflect priviledged users vs no priviledged
user (roles general and product manager recieve product links)
-Manage books page changes to reflect priviledged users. Also the
addition of the image file upload to the server currently the max upload
size is 200kb may change it later. Currently a upload file is not
required to add a new product not sure how everyone feels about that,
would appreciate input.