order requirements
checkout/confirmation/functions - checkout page will check to see if the
customer has a credit card on database and confirmation page will check
to see if shipping address is in the database if it's a shipping order.

inventory - deleted the login check. this was what caused the "you must
login" error to not display. the check is in the header file so this was
redundant.